Rename values config module
